add SpatialRegression package, move related files
